refac: move data handling out of table
